# PoC: Add `rewrite` support for `errorprone.refasterrules`