--- a/android/guava-testlib/src/com/google/common/collect/testing/TestsForSetsInJavaUtil.java
+++ b/android/guava-testlib/src/com/google/common/collect/testing/TestsForSetsInJavaUtil.java
@@ -206,7 +206,7 @@ public class TestsForSetsInJavaUtil {
             new TestEnumSetGenerator() {
               @Override
               public Set<AnEnum> create(AnEnum[] elements) {
-                return (elements.length == 0)
+                return elements.length == 0
                     ? EnumSet.noneOf(AnEnum.class)
                     : EnumSet.copyOf(MinimalCollection.of(elements));
               }