--- a/android/guava-testlib/src/com/google/common/collect/testing/testers/SetHashCodeTester.java
+++ b/android/guava-testlib/src/com/google/common/collect/testing/testers/SetHashCodeTester.java
@@ -41,7 +41,7 @@ public class SetHashCodeTester<E> extends AbstractSetTester<E> {
   public void testHashCode() {
     int expectedHashCode = 0;
     for (E element : getSampleElements()) {
-      expectedHashCode += ((element == null) ? 0 : element.hashCode());
+      expectedHashCode += element == null ? 0 : element.hashCode();
     }
     assertEquals(
         "A Set's hashCode() should be the sum of those of its elements.",
@@ -55,7 +55,7 @@ public class SetHashCodeTester<E> extends AbstractSetTester<E> {
     Collection<E> elements = getSampleElements(getNumElements() - 1);
     int expectedHashCode = 0;
     for (E element : elements) {
-      expectedHashCode += ((element == null) ? 0 : element.hashCode());
+      expectedHashCode += element == null ? 0 : element.hashCode();
     }
 
     elements.add(null);