--- a/android/guava-testlib/src/com/google/common/testing/ArbitraryInstances.java
+++ b/android/guava-testlib/src/com/google/common/testing/ArbitraryInstances.java
@@ -353,7 +353,7 @@ public final class ArbitraryInstances {
     }
     if (type.isEnum()) {
       T[] enumConstants = type.getEnumConstants();
-      return (enumConstants == null || enumConstants.length == 0) ? null : enumConstants[0];
+      return enumConstants == null || enumConstants.length == 0 ? null : enumConstants[0];
     }
     if (type.isArray()) {
       return createEmptyArray(type);