--- a/android/guava/src/com/google/common/base/CaseFormat.java
+++ b/android/guava/src/com/google/common/base/CaseFormat.java
@@ -129,7 +129,7 @@ public enum CaseFormat {
   public final String to(CaseFormat format, String str) {
     checkNotNull(format);
     checkNotNull(str);
-    return (format == this) ? str : convert(format, str);
+    return format == this ? str : convert(format, str);
   }
 
   /** Enum values can override for performance reasons. */
@@ -149,7 +149,7 @@ public enum CaseFormat {
       out.append(format.wordSeparator);
       i = j + wordSeparator.length();
     }
-    return (i == 0)
+    return i == 0
         ? format.normalizeFirstWord(s)
         : requireNonNull(out).append(format.normalizeWord(s.substring(i))).toString();
   }