--- a/android/guava/src/com/google/common/base/internal/Finalizer.java
+++ b/android/guava/src/com/google/common/base/internal/Finalizer.java
@@ -120,7 +120,7 @@ public class Finalizer implements Runnable {
       getBigThreadConstructor();
 
   private static final @Nullable Field inheritableThreadLocals =
-      (bigThreadConstructor == null) ? getInheritableThreadLocalsField() : null;
+      bigThreadConstructor == null ? getInheritableThreadLocalsField() : null;
 
   /** Constructs a new finalizer thread. */
   private Finalizer(