--- a/android/guava/src/com/google/common/cache/LocalCache.java
+++ b/android/guava/src/com/google/common/cache/LocalCache.java
@@ -257,7 +257,7 @@ final class LocalCache<K, V> extends AbstractMap<K, V> implements ConcurrentMap<
 
     removalListener = builder.getRemovalListener();
     removalNotificationQueue =
-        (removalListener == NullListener.INSTANCE)
+        removalListener == NullListener.INSTANCE
             ? LocalCache.discardingQueue()
             : new ConcurrentLinkedQueue<>();
 
@@ -387,7 +387,7 @@ final class LocalCache<K, V> extends AbstractMap<K, V> implements ConcurrentMap<
       @Override
       <K, V> ValueReference<K, V> referenceValue(
           Segment<K, V> segment, ReferenceEntry<K, V> entry, V value, int weight) {
-        return (weight == 1)
+        return weight == 1
             ? new StrongValueReference<K, V>(value)
             : new WeightedStrongValueReference<K, V>(value, weight);
       }
@@ -401,7 +401,7 @@ final class LocalCache<K, V> extends AbstractMap<K, V> implements ConcurrentMap<
       @Override
       <K, V> ValueReference<K, V> referenceValue(
           Segment<K, V> segment, ReferenceEntry<K, V> entry, V value, int weight) {
-        return (weight == 1)
+        return weight == 1
             ? new SoftValueReference<K, V>(segment.valueReferenceQueue, value, entry)
             : new WeightedSoftValueReference<K, V>(
                 segment.valueReferenceQueue, value, entry, weight);
@@ -416,7 +416,7 @@ final class LocalCache<K, V> extends AbstractMap<K, V> implements ConcurrentMap<
       @Override
       <K, V> ValueReference<K, V> referenceValue(
           Segment<K, V> segment, ReferenceEntry<K, V> entry, V value, int weight) {
-        return (weight == 1)
+        return weight == 1
             ? new WeakValueReference<K, V>(segment.valueReferenceQueue, value, entry)
             : new WeightedWeakValueReference<K, V>(
                 segment.valueReferenceQueue, value, entry, weight);
@@ -588,7 +588,7 @@ final class LocalCache<K, V> extends AbstractMap<K, V> implements ConcurrentMap<
     static EntryFactory getFactory(
         Strength keyStrength, boolean usesAccessQueue, boolean usesWriteQueue) {
       int flags =
-          ((keyStrength == Strength.WEAK) ? WEAK_MASK : 0)
+          (keyStrength == Strength.WEAK ? WEAK_MASK : 0)
               | (usesAccessQueue ? ACCESS_MASK : 0)
               | (usesWriteQueue ? WRITE_MASK : 0);
       return factories[flags];
@@ -1679,9 +1679,9 @@ final class LocalCache<K, V> extends AbstractMap<K, V> implements ConcurrentMap<
     // using variant of single-word Wang/Jenkins hash.
     // TODO(kevinb): use Hashing/move this to Hashing?
     h += (h << 15) ^ 0xffffcd7d;
-    h ^= (h >>> 10);
-    h += (h << 3);
-    h ^= (h >>> 6);
+    h ^= h >>> 10;
+    h += h << 3;
+    h ^= h >>> 6;
     h += (h << 2) + (h << 14);
     return h ^ (h >>> 16);
   }
@@ -3041,7 +3041,7 @@ final class LocalCache<K, V> extends AbstractMap<K, V> implements ConcurrentMap<
             newCount = this.count - 1;
             table.set(index, newFirst);
             this.count = newCount; // write-volatile
-            return (cause == RemovalCause.EXPLICIT);
+            return cause == RemovalCause.EXPLICIT;
           }
         }
 
@@ -3084,7 +3084,7 @@ final class LocalCache<K, V> extends AbstractMap<K, V> implements ConcurrentMap<
               ++modCount;
               if (oldValueReference.isActive()) {
                 RemovalCause cause =
-                    (entryValue == null) ? RemovalCause.COLLECTED : RemovalCause.REPLACED;
+                    entryValue == null ? RemovalCause.COLLECTED : RemovalCause.REPLACED;
                 enqueueNotification(key, hash, entryValue, oldValueReference.getWeight(), cause);
                 newCount--;
               }
@@ -3128,7 +3128,7 @@ final class LocalCache<K, V> extends AbstractMap<K, V> implements ConcurrentMap<
                 K key = e.getKey();
                 V value = e.getValueReference().get();
                 RemovalCause cause =
-                    (key == null || value == null) ? RemovalCause.COLLECTED : RemovalCause.EXPLICIT;
+                    key == null || value == null ? RemovalCause.COLLECTED : RemovalCause.EXPLICIT;
                 enqueueNotification(
                     key, e.getHash(), value, e.getValueReference().getWeight(), cause);
               }
@@ -3590,7 +3590,7 @@ final class LocalCache<K, V> extends AbstractMap<K, V> implements ConcurrentMap<
     @Override
     public @Nullable ReferenceEntry<K, V> peek() {
       ReferenceEntry<K, V> next = head.getNextInWriteQueue();
-      return (next == head) ? null : next;
+      return next == head ? null : next;
     }
 
     @Override
@@ -3659,7 +3659,7 @@ final class LocalCache<K, V> extends AbstractMap<K, V> implements ConcurrentMap<
         @Override
         protected @Nullable ReferenceEntry<K, V> computeNext(ReferenceEntry<K, V> previous) {
           ReferenceEntry<K, V> next = previous.getNextInWriteQueue();
-          return (next == head) ? null : next;
+          return next == head ? null : next;
         }
       };
     }
@@ -3730,7 +3730,7 @@ final class LocalCache<K, V> extends AbstractMap<K, V> implements ConcurrentMap<
     @Override
     public @Nullable ReferenceEntry<K, V> peek() {
       ReferenceEntry<K, V> next = head.getNextInAccessQueue();
-      return (next == head) ? null : next;
+      return next == head ? null : next;
     }
 
     @Override
@@ -3799,7 +3799,7 @@ final class LocalCache<K, V> extends AbstractMap<K, V> implements ConcurrentMap<
         @Override
         protected @Nullable ReferenceEntry<K, V> computeNext(ReferenceEntry<K, V> previous) {
           ReferenceEntry<K, V> next = previous.getNextInAccessQueue();
-          return (next == head) ? null : next;
+          return next == head ? null : next;
         }
       };
     }
@@ -3889,7 +3889,7 @@ final class LocalCache<K, V> extends AbstractMap<K, V> implements ConcurrentMap<
   @Override
   public @Nullable V getOrDefault(@Nullable Object key, @Nullable V defaultValue) {
     V result = get(key);
-    return (result != null) ? result : defaultValue;
+    return result != null ? result : defaultValue;
   }
 
   V getOrLoad(K key) throws ExecutionException {
@@ -4179,7 +4179,7 @@ final class LocalCache<K, V> extends AbstractMap<K, V> implements ConcurrentMap<
   public Set<K> keySet() {
     // does not impact recency ordering
     Set<K> ks = keySet;
-    return (ks != null) ? ks : (keySet = new KeySet());
+    return ks != null ? ks : (keySet = new KeySet());
   }
 
   @LazyInit @RetainedWith @Nullable Collection<V> values;
@@ -4188,7 +4188,7 @@ final class LocalCache<K, V> extends AbstractMap<K, V> implements ConcurrentMap<
   public Collection<V> values() {
     // does not impact recency ordering
     Collection<V> vs = values;
-    return (vs != null) ? vs : (values = new Values());
+    return vs != null ? vs : (values = new Values());
   }
 
   @LazyInit @RetainedWith @Nullable Set<Entry<K, V>> entrySet;
@@ -4198,7 +4198,7 @@ final class LocalCache<K, V> extends AbstractMap<K, V> implements ConcurrentMap<
   public Set<Entry<K, V>> entrySet() {
     // does not impact recency ordering
     Set<Entry<K, V>> es = entrySet;
-    return (es != null) ? es : (entrySet = new EntrySet());
+    return es != null ? es : (entrySet = new EntrySet());
   }
 
   // Iterator Support
@@ -4551,7 +4551,7 @@ final class LocalCache<K, V> extends AbstractMap<K, V> implements ConcurrentMap<
       this.weigher = weigher;
       this.concurrencyLevel = concurrencyLevel;
       this.removalListener = removalListener;
-      this.ticker = (ticker == Ticker.systemTicker() || ticker == NULL_TICKER) ? null : ticker;
+      this.ticker = ticker == Ticker.systemTicker() || ticker == NULL_TICKER ? null : ticker;
       this.loader = loader;
     }