--- a/android/guava/src/com/google/common/collect/AbstractBiMap.java
+++ b/android/guava/src/com/google/common/collect/AbstractBiMap.java
@@ -208,7 +208,7 @@ abstract class AbstractBiMap<K extends @Nullable Object, V extends @Nullable Obj
   @Override
   public Set<K> keySet() {
     Set<K> result = keySet;
-    return (result == null) ? keySet = new KeySet() : result;
+    return result == null ? keySet = new KeySet() : result;
   }
 
   @WeakOuter
@@ -257,7 +257,7 @@ abstract class AbstractBiMap<K extends @Nullable Object, V extends @Nullable Obj
      * iteration order so that it is consistent with the forward map.
      */
     Set<V> result = valueSet;
-    return (result == null) ? valueSet = new ValueSet() : result;
+    return result == null ? valueSet = new ValueSet() : result;
   }
 
   @WeakOuter
@@ -296,7 +296,7 @@ abstract class AbstractBiMap<K extends @Nullable Object, V extends @Nullable Obj
   @Override
   public Set<Entry<K, V>> entrySet() {
     Set<Entry<K, V>> result = entrySet;
-    return (result == null) ? entrySet = new EntrySet() : result;
+    return result == null ? entrySet = new EntrySet() : result;
   }
 
   private final class BiMapEntry extends ForwardingMapEntry<K, V> {