--- a/android/guava/src/com/google/common/collect/AbstractMapBasedMultimap.java
+++ b/android/guava/src/com/google/common/collect/AbstractMapBasedMultimap.java
@@ -306,7 +306,7 @@ abstract class AbstractMapBasedMultimap<K extends @Nullable Object, V extends @N
 
   final List<V> wrapList(
       @ParametricNullness K key, List<V> list, @Nullable WrappedCollection ancestor) {
-    return (list instanceof RandomAccess)
+    return list instanceof RandomAccess
         ? new RandomAccessWrappedList(key, list, ancestor)
         : new WrappedList(key, list, ancestor);
   }
@@ -338,7 +338,7 @@ abstract class AbstractMapBasedMultimap<K extends @Nullable Object, V extends @N
       this.key = key;
       this.delegate = delegate;
       this.ancestor = ancestor;
-      this.ancestorDelegate = (ancestor == null) ? null : ancestor.getDelegate();
+      this.ancestorDelegate = ancestor == null ? null : ancestor.getDelegate();
     }
 
     /**
@@ -515,7 +515,7 @@ abstract class AbstractMapBasedMultimap<K extends @Nullable Object, V extends @N
       boolean changed = delegate.addAll(collection);
       if (changed) {
         int newSize = delegate.size();
-        totalSize += (newSize - oldSize);
+        totalSize += newSize - oldSize;
         if (oldSize == 0) {
           addToMap();
         }
@@ -566,7 +566,7 @@ abstract class AbstractMapBasedMultimap<K extends @Nullable Object, V extends @N
       boolean changed = delegate.removeAll(c);
       if (changed) {
         int newSize = delegate.size();
-        totalSize += (newSize - oldSize);
+        totalSize += newSize - oldSize;
         removeIfEmpty();
       }
       return changed;
@@ -579,7 +579,7 @@ abstract class AbstractMapBasedMultimap<K extends @Nullable Object, V extends @N
       boolean changed = delegate.retainAll(c);
       if (changed) {
         int newSize = delegate.size();
-        totalSize += (newSize - oldSize);
+        totalSize += newSize - oldSize;
         removeIfEmpty();
       }
       return changed;
@@ -588,7 +588,7 @@ abstract class AbstractMapBasedMultimap<K extends @Nullable Object, V extends @N
 
   private static <E extends @Nullable Object> Iterator<E> iteratorOrListIterator(
       Collection<E> collection) {
-    return (collection instanceof List)
+    return collection instanceof List
         ? ((List<E>) collection).listIterator()
         : collection.iterator();
   }
@@ -613,7 +613,7 @@ abstract class AbstractMapBasedMultimap<K extends @Nullable Object, V extends @N
       boolean changed = Sets.removeAllImpl((Set<V>) delegate, c);
       if (changed) {
         int newSize = delegate.size();
-        totalSize += (newSize - oldSize);
+        totalSize += newSize - oldSize;
         removeIfEmpty();
       }
       return changed;
@@ -657,7 +657,7 @@ abstract class AbstractMapBasedMultimap<K extends @Nullable Object, V extends @N
       return new WrappedSortedSet(
           getKey(),
           getSortedSetDelegate().headSet(toElement),
-          (getAncestor() == null) ? this : getAncestor());
+          getAncestor() == null ? this : getAncestor());
     }
 
     @Override
@@ -666,7 +666,7 @@ abstract class AbstractMapBasedMultimap<K extends @Nullable Object, V extends @N
       return new WrappedSortedSet(
           getKey(),
           getSortedSetDelegate().subSet(fromElement, toElement),
-          (getAncestor() == null) ? this : getAncestor());
+          getAncestor() == null ? this : getAncestor());
     }
 
     @Override
@@ -675,7 +675,7 @@ abstract class AbstractMapBasedMultimap<K extends @Nullable Object, V extends @N
       return new WrappedSortedSet(
           getKey(),
           getSortedSetDelegate().tailSet(fromElement),
-          (getAncestor() == null) ? this : getAncestor());
+          getAncestor() == null ? this : getAncestor());
     }
   }
 
@@ -722,7 +722,7 @@ abstract class AbstractMapBasedMultimap<K extends @Nullable Object, V extends @N
     }
 
     private NavigableSet<V> wrap(NavigableSet<V> wrapped) {
-      return new WrappedNavigableSet(key, wrapped, (getAncestor() == null) ? this : getAncestor());
+      return new WrappedNavigableSet(key, wrapped, getAncestor() == null ? this : getAncestor());
     }
 
     @Override
@@ -776,7 +776,7 @@ abstract class AbstractMapBasedMultimap<K extends @Nullable Object, V extends @N
       boolean changed = getListDelegate().addAll(index, c);
       if (changed) {
         int newSize = getDelegate().size();
-        totalSize += (newSize - oldSize);
+        totalSize += newSize - oldSize;
         if (oldSize == 0) {
           addToMap();
         }
@@ -849,7 +849,7 @@ abstract class AbstractMapBasedMultimap<K extends @Nullable Object, V extends @N
       return wrapList(
           getKey(),
           getListDelegate().subList(fromIndex, toIndex),
-          (getAncestor() == null) ? this : getAncestor());
+          getAncestor() == null ? this : getAncestor());
     }
 
     /** ListIterator decorator. */
@@ -1483,7 +1483,7 @@ abstract class AbstractMapBasedMultimap<K extends @Nullable Object, V extends @N
     @Override
     public SortedSet<K> keySet() {
       SortedSet<K> result = sortedKeySet;
-      return (result == null) ? sortedKeySet = createKeySet() : result;
+      return result == null ? sortedKeySet = createKeySet() : result;
     }
 
     @Override
@@ -1506,7 +1506,7 @@ abstract class AbstractMapBasedMultimap<K extends @Nullable Object, V extends @N
     @Override
     public @Nullable Entry<K, Collection<V>> lowerEntry(@ParametricNullness K key) {
       Entry<K, Collection<V>> entry = sortedMap().lowerEntry(key);
-      return (entry == null) ? null : wrapEntry(entry);
+      return entry == null ? null : wrapEntry(entry);
     }
 
     @Override
@@ -1517,7 +1517,7 @@ abstract class AbstractMapBasedMultimap<K extends @Nullable Object, V extends @N
     @Override
     public @Nullable Entry<K, Collection<V>> floorEntry(@ParametricNullness K key) {
       Entry<K, Collection<V>> entry = sortedMap().floorEntry(key);
-      return (entry == null) ? null : wrapEntry(entry);
+      return entry == null ? null : wrapEntry(entry);
     }
 
     @Override
@@ -1528,7 +1528,7 @@ abstract class AbstractMapBasedMultimap<K extends @Nullable Object, V extends @N
     @Override
     public @Nullable Entry<K, Collection<V>> ceilingEntry(@ParametricNullness K key) {
       Entry<K, Collection<V>> entry = sortedMap().ceilingEntry(key);
-      return (entry == null) ? null : wrapEntry(entry);
+      return entry == null ? null : wrapEntry(entry);
     }
 
     @Override
@@ -1539,7 +1539,7 @@ abstract class AbstractMapBasedMultimap<K extends @Nullable Object, V extends @N
     @Override
     public @Nullable Entry<K, Collection<V>> higherEntry(@ParametricNullness K key) {
       Entry<K, Collection<V>> entry = sortedMap().higherEntry(key);
-      return (entry == null) ? null : wrapEntry(entry);
+      return entry == null ? null : wrapEntry(entry);
     }
 
     @Override
@@ -1550,13 +1550,13 @@ abstract class AbstractMapBasedMultimap<K extends @Nullable Object, V extends @N
     @Override
     public @Nullable Entry<K, Collection<V>> firstEntry() {
       Entry<K, Collection<V>> entry = sortedMap().firstEntry();
-      return (entry == null) ? null : wrapEntry(entry);
+      return entry == null ? null : wrapEntry(entry);
     }
 
     @Override
     public @Nullable Entry<K, Collection<V>> lastEntry() {
       Entry<K, Collection<V>> entry = sortedMap().lastEntry();
-      return (entry == null) ? null : wrapEntry(entry);
+      return entry == null ? null : wrapEntry(entry);
     }
 
     @Override