--- a/android/guava/src/com/google/common/collect/AbstractSortedMultiset.java
+++ b/android/guava/src/com/google/common/collect/AbstractSortedMultiset.java
@@ -122,7 +122,7 @@ abstract class AbstractSortedMultiset<E extends @Nullable Object> extends Abstra
   @Override
   public SortedMultiset<E> descendingMultiset() {
     SortedMultiset<E> result = descendingMultiset;
-    return (result == null) ? descendingMultiset = createDescendingMultiset() : result;
+    return result == null ? descendingMultiset = createDescendingMultiset() : result;
   }
 
   SortedMultiset<E> createDescendingMultiset() {