--- a/android/guava/src/com/google/common/collect/CollectCollectors.java
+++ b/android/guava/src/com/google/common/collect/CollectCollectors.java
@@ -339,7 +339,7 @@ final class CollectCollectors {
     }
 
     ImmutableMap<K, V> toImmutableMap() {
-      return (map == null) ? ImmutableMap.<K, V>of() : ImmutableEnumMap.asImmutable(map);
+      return map == null ? ImmutableMap.<K, V>of() : ImmutableEnumMap.asImmutable(map);
     }
   }