```
[WARNING] Changes have been made to guava-gwt/src-super/com/google/common/cache/super/com/google/common/cache/LocalCache.java by:
[WARNING]     org.openrewrite.staticanalysis.MissingOverrideAnnotation
[WARNING] Please review and commit the results.
[WARNING] Estimate time saved: 2h 30m
[INFO] ------------------------------------------------------------------------
[INFO] Reactor Summary for Guava Maven Parent 999.0.0-HEAD-jre-SNAPSHOT:
[INFO] 
[INFO] Guava Maven Parent ................................. SUCCESS [  0.270 s]
[INFO] Guava: Google Core Libraries for Java .............. SUCCESS [  1.327 s]
[INFO] Guava BOM .......................................... SUCCESS [  0.028 s]
[INFO] Guava Testing Library .............................. SUCCESS [  0.289 s]
[INFO] Guava Unit Tests ................................... SUCCESS [  0.143 s]
[INFO] Guava GWT compatible libs .......................... SUCCESS [ 43.500 s]
[INFO] ------------------------------------------------------------------------
[INFO] BUILD SUCCESS
[INFO] ------------------------------------------------------------------------
[INFO] Total time:  45.870 s
[INFO] Finished at: 2025-08-06T11:50:37+02:00
[INFO] ------------------------------------------------------------------------
```