@Pankraz76 2025-08-09
Thanks for the response.

> integrate things fully

You mean the actual runtime?

To make this opt in and still last, we could simply disable by default the execution via switch `rewrite.skip=true`.
 
---

Just discovered you are (of course) a committer to `error-prone` as well, so im wondering why are you hesistating?
Im sure you have good intend and reasoning to do so.

Despite the large time consumption, its generally nice to fixes instead of dealing with them bugs manually, right?