@Pankraz76 2025-08-09
Hi,

might consider these changes suggested by `error.prone`. Then talk about possible CI integration depending on the actual time spend. Just a few minuets might be accepbtable as positive tradeoff for the bugs fixed.

Thanks.

- https://github.com/google/guava/pull/7930